add empty directories analyze, e2etest/test_report and utest/coverage_report
The analyze directory is used to put in account statement files,
which should be processd by the bank_account_monitor.
We only want to track the directory and not the content in git.

The e2etest/test_report/ directory is for storing e2e test reports.
The reports itself should not be tracked either.

The same goes for the utest/coverage_report/ directory which contain
coverage report files after executing the unit tests