refactor SKassenTransactionProvider and VrBankTransactionProvider
First, we removed code duplication in the constructor by moving
the file check to the base class.

Secondly, we introduced the parameter --disable-error-code=method-assign when calling
mypy in ./run_utest.sh.
Unfortunately we have to do that because otherwise mypy always complains with
the following error message when we use MagicMock in a test setup:
error: Cannot assign to a method.

Using MagicMock in a setup function sometimes helps us
to make the test arrange fild smaller.