Enable savings presenter
now it is possible to generate despite the general csv output,
a "savings" pie chart which depicts the distribution of incommings and debits.
The savings pie chart is runnable as standalone as well.
We achieve this by making the input parameter csv_output_file
and savings optional.
But it is mandatory to pass at least one of the optional parameters.
In case of a losing situation we also print pie chart. But now
the pie chart shows only the debits and the losing, what is in that
case not a distribution.