refactor: apply single responsible rule to PresenterFactory
The former PresenterFactory violated the single responsible rule,
because there were two reasons to change.
The first reason was adding a new presenter (main reason) and a second
unintended reason was the plotter data csv parser.
The plotter data in cvs format is ui driven. Changes in the format of
the plotter data in the ui also demands changes in the PresenterFactory.

We introduced a new business logic data class for the PresenterFactory