refactor presenter selection via new config concept
the presenter config is now passed to the PresenterFactory via
a dictionary.
This will allow us in the future to combine several
presenter without changing the code. Also in case we have to add a new
presenter we only have to change the ConfigReader and add a entry to the
PresenterFactory. Former, there where two place to adjust in the
PresenterFactory. One for introducing the new presenter and one to add
this presenter to the MultiPresenter input list.