presenting the savings works now in the sdk
Previously we called pyplot.show, which opened a pop-up window
in which the savings were presented.
Pop-up windows are not possible within a docker container.
That was also the reason why there were no end to end tests for
presenting the savings.

Now we let pyplot store the savings image as png.
Therefore we had to introduce a new field in SavingsPresConfig that let
us configure where the savings image should be stored.