Add new savings and losing png images
We faced the issue that after installing a newer
version of matplotlib the fallowing  e2e tests are
failing:

Store calculated asset savings as pie chart in png
image
Store calculated asset losings as pie chart in png
image
Sort VR bank account statement into csv file and
show savings

The content of the PNG images generated with
matplotlib remains unchanged, but the binary
file itself is different. The e2e test compare
only the binary file.
For the moment we just
update the png image for the newer matplotlib.

The next step will be to use fixed package
versions for within the docker image